fixed auto_clim for current crowding visualization, refactored some helper functions
