+++ tests/palace/test_results.py
@@ -1,4 +1,8 @@
-"""Tests for gsim.palace.results — S-parameter loading with port-name mapping."""
+"""Tests for gsim.palace.results and gsim.palace.fields.
+
+Covers S-parameter loading with port-name mapping (results) and the
+NaN-free field-visualization routines (fields).
+"""
 
 from __future__ import annotations
 
@@ -252,3 +256,422 @@ class TestSParamsSaveLoad:
     def test_from_file_missing_raises(self, tmp_path: Path) -> None:
         with pytest.raises(FileNotFoundError):
             SParams.from_file(tmp_path / "nonexistent.npz")
+
+
+# ---------------------------------------------------------------------------
+# Field-visualization tests (gsim.palace.fields) — NaN-free direct mesh rendering
+# ---------------------------------------------------------------------------
+
+
+def _make_triangle_mesh() -> tuple:
+    """Create a tiny PyVista triangle mesh with a vector field + attribute cell data.
+
+    Returns (mesh, vector_field_name, attribute_field_name).
+    """
+    import pyvista as pv
+
+    # Two triangles forming a square
+    # PyVista 0.46+ requires flat connectivity
+    points = np.array(
+        [
+            [0.0, 0.0, 0.0],  # 0
+            [1.0, 0.0, 0.0],  # 1
+            [1.0, 1.0, 0.0],  # 2
+            [0.0, 1.0, 0.0],  # 3
+        ],
+        dtype=float,
+    )
+    # Flat connectivity: for each cell: [n_nodes, id0, id1, id2]
+    cells = np.array([3, 0, 1, 2, 3, 0, 2, 3], dtype=np.int64)
+    celltypes = np.array([pv.CellType.TRIANGLE, pv.CellType.TRIANGLE], dtype=np.uint8)
+
+    mesh = pv.UnstructuredGrid(cells, celltypes, points)
+
+    # Vector point data (3-component)
+    mesh.point_data["E"] = np.array(
+        [
+            [0.0, 0.0, 1.0],
+            [1.0, 0.0, 0.0],
+            [0.0, 1.0, 0.0],
+            [0.5, 0.5, 0.5],
+        ],
+        dtype=float,
+    )
+
+    # Attribute cell data
+    mesh.cell_data["attribute"] = np.array([1, 2], dtype=int)
+
+    return mesh, "E", "attribute"
+
+
+class TestActivateVectorComponent:
+    def test_magnitude(self):
+        from gsim.palace.fields import activate_vector_component
+
+        mesh, field, _ = _make_triangle_mesh()
+        name = activate_vector_component(mesh, field, component="mag")
+        assert name == "E_mag"
+        assert "E_mag" in mesh.point_data
+        vals = mesh.point_data["E_mag"]
+        assert vals.shape == (4,)
+        assert np.isfinite(vals).all()
+
+    def test_x_component(self):
+        from gsim.palace.fields import activate_vector_component
+
+        mesh, field, _ = _make_triangle_mesh()
+        name = activate_vector_component(mesh, field, component="x")
+        assert name == "E_x"
+        assert "E_x" in mesh.point_data
+
+    def test_custom_output_name(self):
+        from gsim.palace.fields import activate_vector_component
+
+        mesh, field, _ = _make_triangle_mesh()
+        name = activate_vector_component(mesh, field, component="mag", output_name="magnitude")
+        assert name == "magnitude"
+        assert "magnitude" in mesh.point_data
+
+    def test_missing_field(self):
+        from gsim.palace.fields import activate_vector_component
+
+        mesh, _, _ = _make_triangle_mesh()
+        with pytest.raises(KeyError, match="not found"):
+            activate_vector_component(mesh, "nonexistent", component="mag")
+
+    def test_invalid_component(self):
+        from gsim.palace.fields import activate_vector_component
+
+        mesh, field, _ = _make_triangle_mesh()
+        with pytest.raises(ValueError, match="component must be one of"):
+            activate_vector_component(mesh, field, component="w")
+
+
+class TestExtractBoundaryCells:
+    def test_basic_extraction(self):
+        from gsim.palace.fields import extract_boundary_cells
+
+        mesh, _, _ = _make_triangle_mesh()
+        subset = extract_boundary_cells(mesh, [1])
+        assert subset.n_cells == 1
+        assert subset.n_points > 0
+
+    def test_multiple_attributes(self):
+        from gsim.palace.fields import extract_boundary_cells
+
+        mesh, _, _ = _make_triangle_mesh()
+        subset = extract_boundary_cells(mesh, {1, 2})
+        assert subset.n_cells == 2
+
+    def test_no_match(self):
+        from gsim.palace.fields import extract_boundary_cells
+
+        mesh, _, _ = _make_triangle_mesh()
+        with pytest.raises(ValueError, match="No cells matched"):
+            extract_boundary_cells(mesh, [99])
+
+    def test_empty_attributes(self):
+        from gsim.palace.fields import extract_boundary_cells
+
+        mesh, _, _ = _make_triangle_mesh()
+        with pytest.raises(ValueError, match="No attributes provided"):
+            extract_boundary_cells(mesh, [])
+
+
+class TestResolveScalarField:
+    def test_scalar_direct(self):
+        from gsim.palace.fields import resolve_scalar_field
+
+        mesh, _, _ = _make_triangle_mesh()
+        mesh.point_data["my_scalar"] = np.ones(mesh.n_points)
+        name = resolve_scalar_field(mesh, scalar_field="my_scalar")
+        assert name == "my_scalar"
+
+    def test_vector_field(self):
+        from gsim.palace.fields import resolve_scalar_field
+
+        mesh, field, _ = _make_triangle_mesh()
+        name = resolve_scalar_field(mesh, vector_field=field, component="mag")
+        assert name == "E_mag"
+        assert "E_mag" in mesh.point_data
+
+    def test_both_provided(self):
+        from gsim.palace.fields import resolve_scalar_field
+
+        mesh, _, _ = _make_triangle_mesh()
+        with pytest.raises(ValueError, match="not both"):
+            resolve_scalar_field(mesh, scalar_field="a", vector_field="b")
+
+    def test_neither_provided(self):
+        from gsim.palace.fields import resolve_scalar_field
+
+        mesh, _, _ = _make_triangle_mesh()
+        with pytest.raises(ValueError, match="Provide one of"):
+            resolve_scalar_field(mesh)
+
+
+class TestPlotBoundaryField:
+    """Smoke tests — create plotter but don't show."""
+
+    def test_vector_field_render(self):
+        from gsim.palace.fields import BoundaryFieldData, plot_boundary_field
+
+        mesh, field, _ = _make_triangle_mesh()
+        bnd = BoundaryFieldData(
+            mesh=mesh,
+            dataset_name="test",
+            step_index=0,
+            timestep=0.0,
+            selected_attributes=(1, 2),
+        )
+        pl = plot_boundary_field(bnd, vector_field=field, component="mag", off_screen=True)
+        assert pl is not None
+        pl.close()
+
+    def test_scalar_field_render(self):
+        from gsim.palace.fields import BoundaryFieldData, plot_boundary_field
+
+        mesh, _, _ = _make_triangle_mesh()
+        mesh.point_data["my_scalar"] = np.ones(mesh.n_points)
+        bnd = BoundaryFieldData(
+            mesh=mesh,
+            dataset_name="test",
+            step_index=0,
+            timestep=0.0,
+            selected_attributes=(1, 2),
+        )
+        pl = plot_boundary_field(bnd, scalar_field="my_scalar", off_screen=True)
+        assert pl is not None
+        pl.close()
+
+    def test_log_scale_positive(self):
+        from gsim.palace.fields import BoundaryFieldData, plot_boundary_field
+
+        mesh, field, _ = _make_triangle_mesh()
+        # Ensure all values positive for log scale
+        mesh.point_data[field] = np.abs(mesh.point_data[field]) + 0.1
+        bnd = BoundaryFieldData(
+            mesh=mesh,
+            dataset_name="test",
+            step_index=0,
+            timestep=0.0,
+            selected_attributes=(1, 2),
+        )
+        pl = plot_boundary_field(bnd, vector_field=field, component="mag", log_scale=True, off_screen=True)
+        assert pl is not None
+        pl.close()
+
+    def test_no_nan_after_magnitude(self):
+        """Core assertion: computed magnitude must have 0 NaN."""
+        from gsim.palace.fields import BoundaryFieldData, activate_vector_component
+
+        mesh, field, _ = _make_triangle_mesh()
+        scalar_name = activate_vector_component(mesh, field, component="mag")
+        assert np.isnan(mesh.point_data[scalar_name]).sum() == 0
+
+    def test_log_scale_with_nonpositive(self):
+        """Log scale with some zero values must clamp correctly, no errors."""
+        from gsim.palace.fields import BoundaryFieldData, plot_boundary_field
+
+        mesh, field, _ = _make_triangle_mesh()
+        # Set one point to zero
+        mesh.point_data[field][0] = [0.0, 0.0, 0.0]
+        bnd = BoundaryFieldData(
+            mesh=mesh,
+            dataset_name="test",
+            step_index=0,
+            timestep=0.0,
+            selected_attributes=(1, 2),
+        )
+        pl = plot_boundary_field(bnd, vector_field=field, component="mag", log_scale=True, off_screen=True)
+        pl.close()
+
+    def test_auto_clim_percentile(self):
+        """Auto clim uses 98th percentile so peaks don't wash out the bulk."""
+        from gsim.palace.fields import _auto_clim
+
+        # bulk ~1, a few peaks ~1000 (current crowding)
+        vals = np.concatenate([np.full(900, 1.0), np.full(100, 1000.0)])
+        vmin, vmax = _auto_clim(vals, log_scale=False)
+        assert vmin == 0.0
+        # 98th percentile of this distribution is 1.0 (peaks are top 2%)
+        assert vmax == pytest.approx(1.0)
+
+    def test_auto_clim_log(self):
+        from gsim.palace.fields import _auto_clim
+
+        vals = np.concatenate([np.full(900, 1.0), np.full(100, 1000.0)])
+        vmin, vmax = _auto_clim(vals, log_scale=True)
+        assert vmin > 0.0
+        assert vmax > vmin
+
+    def test_auto_clim_signed_symmetric(self):
+        """Signed data gets symmetric limits centered at zero."""
+        from gsim.palace.fields import _auto_clim
+
+        # E_y-like data: both polarities, peaks at edges
+        vals = np.concatenate([np.full(400, -100.0), np.full(400, 100.0),
+                               np.linspace(-500, 500, 200)])
+        vmin, vmax = _auto_clim(vals, signed=True)
+        assert vmin < 0.0 < vmax
+        assert vmin == pytest.approx(-vmax)  # symmetric
+
+    def test_auto_clim_unsigned_clamps_to_zero(self):
+        """Unsigned data (magnitudes) clamps vmin to 0."""
+        from gsim.palace.fields import _auto_clim
+
+        vals = np.abs(np.concatenate([np.full(400, -100.0), np.full(100, 500.0)]))
+        vmin, vmax = _auto_clim(vals, signed=False)
+        assert vmin == 0.0
+        assert vmax > 0.0
+
+
+class TestSelectorContext:
+    def test_build_from_dict(self):
+        from gsim.palace.fields import build_selector_context
+
+        config = {"Boundaries": {"PEC": {"Attributes": [1, 2]}}}
+        pg_map = {"top_conductor": 1, "ground_plane": 2}
+        ctx = build_selector_context(config, pg_map)
+        assert ctx.pg_map["top_conductor"] == 1
+        assert ctx.boundaries_by_type["PEC"] == (1, 2)
+
+    def test_resolve_entity_attributes(self):
+        from gsim.palace.fields import (
+            SelectorContext,
+            resolve_entity_attributes,
+        )
+
+        ctx = SelectorContext(pg_map={"a": 10, "b": 20}, boundaries_by_type={})
+        attrs = resolve_entity_attributes(["a", "b"], ctx)
+        assert attrs == (10, 20)
+
+    def test_resolve_unknown_entity(self):
+        from gsim.palace.fields import (
+            SelectorContext,
+            resolve_entity_attributes,
+        )
+
+        ctx = SelectorContext(pg_map={"a": 10}, boundaries_by_type={})
+        with pytest.raises(KeyError, match="Unknown"):
+            resolve_entity_attributes("missing", ctx)
+
+
+class TestResolveBoundaryType:
+    def test_resolve_boundary_type(self):
+        from gsim.palace.fields import (
+            SelectorContext,
+            resolve_boundary_type_attributes,
+        )
+
+        ctx = SelectorContext(
+            pg_map={},
+            boundaries_by_type={"PEC": (1, 2)},
+        )
+        assert resolve_boundary_type_attributes("PEC", ctx) == (1, 2)
+
+    def test_unknown_boundary_type(self):
+        from gsim.palace.fields import (
+            SelectorContext,
+            resolve_boundary_type_attributes,
+        )
+
+        ctx = SelectorContext(pg_map={}, boundaries_by_type={})
+        with pytest.raises(KeyError, match="not found"):
+            resolve_boundary_type_attributes("ABC", ctx)
+
+
+class TestPlotVolumeSlice:
+    def test_slice_render(self):
+        from gsim.palace.fields import VolumeFieldData, plot_volume_slice
+
+        mesh, field, _ = _make_triangle_mesh()
+        vol = VolumeFieldData(mesh=mesh, dataset_name="test", step_index=0, timestep=0.0)
+        pl = plot_volume_slice(vol, vector_field=field, axis="z", value=0.0, off_screen=True)
+        pl.close()
+
+
+class TestExtractAxisSlice:
+    def test_extract(self):
+        from gsim.palace.fields import VolumeFieldData, extract_axis_slice
+
+        mesh, _, _ = _make_triangle_mesh()
+        vol = VolumeFieldData(mesh=mesh, dataset_name="test", step_index=0, timestep=0.0)
+        sl = extract_axis_slice(vol, axis="z", value=0.0)
+        assert sl.n_points > 0
+
+    def test_invalid_axis(self):
+        from gsim.palace.fields import VolumeFieldData, extract_axis_slice
+
+        mesh, _, _ = _make_triangle_mesh()
+        vol = VolumeFieldData(mesh=mesh, dataset_name="test", step_index=0, timestep=0.0)
+        with pytest.raises(ValueError, match="axis must be one of"):
+            extract_axis_slice(vol, axis="w", value=0.0)
+
+
+class TestLoadFieldContext:
+    """Tests for the load_field_context convenience helper."""
+
+    def test_loads_meshes_and_context(self, tmp_path, monkeypatch):
+        from types import SimpleNamespace
+
+        from gsim.palace import fields as fields_mod
+
+        # Build a fake simulation directory layout:
+        #   <sim_dir>/config.json
+        #   <sim_dir>/palace.msh
+        #   <sim_dir>/output/palace/port-S.csv   <- results_dir
+        sim_dir = tmp_path / "sim"
+        results_dir = sim_dir / "output" / "palace"
+        results_dir.mkdir(parents=True)
+        (results_dir / "port-S.csv").write_text("f (GHz)\n1.0\n")
+        (sim_dir / "palace.msh").write_text("mesh")
+        (sim_dir / "config.json").write_text(
+            json.dumps({"Boundaries": {"PEC": {"Attributes": [1, 2]}}})
+        )
+
+        # Fake meshio.read returning an object with field_data
+        fake_mesh = SimpleNamespace(field_data={"topmetal2_xy": (5, 2), "ground": (6, 2)})
+        import meshio as _real_meshio
+
+        monkeypatch.setattr(_real_meshio, "read", lambda _: fake_mesh)
+
+        # Fake load_fields returning distinct volume/boundary meshes
+        fake_vol = SimpleNamespace(n_points=100, n_cells=10, _kind="vol")
+        fake_bnd = SimpleNamespace(n_points=50, n_cells=5, _kind="bnd")
+
+        def fake_load_fields(source, *, excitation=1, cycle=None, boundary=False):
+            return fake_bnd if boundary else fake_vol
+
+        monkeypatch.setattr("gsim.palace.results.load_fields", fake_load_fields)
+
+        vol, bnd, ctx, pg_map = fields_mod.load_field_context(results_dir, excitation=1)
+
+        assert vol is fake_vol
+        assert bnd is fake_bnd
+        assert pg_map == {"topmetal2_xy": 5, "ground": 6}
+        assert ctx.boundaries_by_type["PEC"] == (1, 2)
+
+    def test_missing_mesh_raises(self, tmp_path):
+        from gsim.palace import fields as fields_mod
+
+        sim_dir = tmp_path / "sim"
+        results_dir = sim_dir / "output" / "palace"
+        results_dir.mkdir(parents=True)
+        (results_dir / "port-S.csv").write_text("f (GHz)\n1.0\n")
+        (sim_dir / "config.json").write_text(json.dumps({}))
+
+        with pytest.raises(FileNotFoundError, match="Mesh file not found"):
+            fields_mod.load_field_context(results_dir)
+
+    def test_missing_config_raises(self, tmp_path):
+        from gsim.palace import fields as fields_mod
+
+        sim_dir = tmp_path / "sim"
+        results_dir = sim_dir / "output" / "palace"
+        results_dir.mkdir(parents=True)
+        (results_dir / "port-S.csv").write_text("f (GHz)\n1.0\n")
+        (sim_dir / "palace.msh").write_text("mesh")
+
+        with pytest.raises(FileNotFoundError, match="Palace config not found"):
+            fields_mod.load_field_context(results_dir)
